Städat lite, bytt ut tempdata med fel mot den utan
