--- AverageTempDayInside.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-
-namespace Arch_VaderData
-{
-    //internal class AverageTempDayInside
-    //{
-    //    private static string tempData = "tempdata5-med fel.txt";
-
-
-    //    public static void AverageTempInsideDesc(string location)
-    //    {
-    //        double averageTemp = 0;
-    //        double averageHumidity = 0;
-
-    //        using (StreamReader reader = new StreamReader(tempData))
-    //        {
-    //            while (true)
-    //            {
-    //                string dateToCheck = "^";
-
-
-    //                RegexOptions options = RegexOptions.Multiline;
-    //                string fullData = reader.ReadToEnd();
-    //                Regex data = new Regex(dateToCheck, options);
-    //                MatchCollection matches = data.Matches(fullData);
-
-
-
-        //}
-}
-
